crmgym/src/main/java/com/epam/crmgym/exception/BindingResultError.java
package com.epam.crmgym.exception;


import org.springframework.stereotype.Component;
import org.springframework.validation.BindingResult;

import java.util.List;
import java.util.stream.Collectors;

@Component
public class BindingResultError {
    public List<String> handleBindingResultErrors(BindingResult bindingResult) {
        return bindingResult.getFieldErrors().stream()
                .map(error -> String.format("%s: %s", error.getField(), error.getDefaultMessage()))
                .collect(Collectors.toList());
    }
}
